pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imagegen-server"
version = "0.1.0"
description = "Queued image generation service with multi-key worker balancing."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115,<1",
  "openai>=1.107.0,<2",
  "pillow>=11,<12",
  "uvicorn[standard]>=0.30,<1",
  "python-multipart>=0.0.9,<1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
imagegen_server = ["web/*.html", "web/*.css", "web/*.js"]
